+++ app/src/main/java/com/xizz/greeting/AppComponent.kt
@@ -1,4 +1,4 @@
-package com.xizz.anvilsetup
+package com.xizz.greeting
 
 import com.squareup.anvil.annotations.MergeComponent
