+++ app/src/main/java/com/xizz/greeting/AppScope.kt
@@ -1,3 +1,3 @@
-package com.xizz.anvilsetup
+package com.xizz.greeting
 
 abstract class AppScope private constructor()
