+++ app/src/main/java/com/xizz/greeting/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.xizz.anvilsetup.ui.theme
+package com.xizz.greeting.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
